requirements.txt
# Blue Pharma Trading PLC Telegram Bot Requirements

# Core Telegram Bot
python-telegram-bot==22.3

# Environment Management
python-dotenv==1.0.1

# Date and Time
python-dateutil==2.9.0

# Logging
colorlog==6.8.2

# Input Validation
validators==0.22.0

# Utilities
typing-extensions==4.12.2

# Analytics Dependencies
pandas==2.2.2
openpyxl==3.1.5

# Development/Testing (optional)
pytest==8.2.2
pytest-asyncio==0.23.8
